Update getting_started.rst
Updates to the Getting Started.  Re-organized to provide concept and process info at the beginning.
Followed by "Signup for ObjectRocket" and "Test your database connection".

Philosophy:  The signup wizard is by definition a tutorial for signing up, so we probably don't need to go through the wizard steps in the Getting Started.  I revised this to provide people with the information that they need before signing up.  Then, link them to the wizard.

I haven't gone farther here, but I believe it would be helpful to extend this Getting Started to include a sample of connecting to the database, a troubleshooting section to list common issues, and a next steps.

Also, maybe the Getting Started should provide some overveiw of the Dashboard functionality?