Update getting_started.rst
Made substantial revisions to the Getting Started Guide. Philosophy:  Tell user what is required, explain important concepts and processes that might be new, and provide high-level overview of initial set up and configuration process.

Because you have a wizard, the documentation doesn't need to provide detailed information about the steps.  If the wizard isn't intuitive, then add text to wizard UI or improve process. 

One thing that can be really confusing for new users is combining explanations of concepts and processes within the procedure to get the task done.